more edits

+++ routeTemp.js
@@ -42,7 +42,7 @@ router.get('/categories', (req, res) => {
             console.log('Error finding categories');
             throw err;
         }
-        res.send(result);
+        res.render('categories.ejs', { categories: result });
     })
 })
 /* Scientific discipline detail page */
@@ -53,7 +53,7 @@ router.get('/categories/:scientific_discipline', (req, res) => {
             console.log('Error finding categories');
             throw err;
         }
-        res.send(result);
+        res.render('category.ejs', { category: result });
     })
 })
 /* Tags index page */
@@ -63,7 +63,7 @@ router.get('/keywords', (req, res) => {
             console.log('Error finding tags');
             throw err;
         }
-        res.send(result);
+        res.res.render('keywords.ejs', { keywords: result });
     })
 })
 /*Tag page*/
@@ -74,7 +74,7 @@ router.get('/keywords/:tag', (req, res) => {
             console.log('Error finding tags');
             throw err;
         }
-        res.send(result);
+        res.render('keyword.ejs', { keyword: result });
     })
 })
 /* Podcast page */
@@ -85,7 +85,7 @@ router.get('/podcasts/:podcast_title', (req, res) => {
             console.log('Error finding podcasts');
             throw err;
         }
-        res.send(result);
+        res.render('podcast.ejs', { podcast: result });
     })
 })
 /* Authentication */
@@ -105,67 +105,93 @@ router.get('/authors/create', (req, res) => {
 /*User/author profile */
 router.get('/users/:first_name_last_name', (req, res) => {
     var name = req.params.first_name_last_name;
-    name.split('-')
-    db.collection('users').find({ fname: name[0] }, function (err, result) {
-        if (err) {
-            console.log('Error finding users');
-            throw err;
-        }
-        forEach(result, function (user) {
-            if (user['lname'] == name[1])
-                res.send(user);
+    name.split(/-.*_/)
+    if (name.length > 2) {
+        db.collection('users').find({ fname: name[0], lname: name[1] }, function (err, result) {
+            if (err) {
+                console.log('Error finding users');
+                throw err;
+            }
+            res.render('profile.ejs', { user: result[name[2]] });
         })
-    })
+    }
+    else {
+        db.collection('users').findOne({ fname: name[0], lname: name[1] }, function (err, result) {
+            if (err) {
+                console.log('Error finding users');
+                throw err;
+            }
+            res.render('profile.ejs', { user: result[0] });
+        })
+    }
 })
 /* User/author account */
 router.get('/account', (req, res) => {
     //if the session is logged in, render the account page
     var session = req.session;
+    var user = session.username;
+    user.split(/-.*_/)
     if (session != null) {
-        db.collection('users').findOne({ username: session.username }, function (err, result) {
-            res.render('./pages/profile.ejs', { user: result });
-        })
+        if (user.length == 2) {
+            db.collection('users').findOne({ fname: user[0], lname: user[1] }, function (err, result) {
+                res.render('./pages/profile.ejs', { user: result });
+            })
+        }
+        else {
+            db.collection('users').find({ fname: user[0], lname: user[1] }, function (err, result) {
+                res.render('./pages/profile.ejs', { user: result[user[2]] });
+            })
+        }
     }
 })
 /* Edit/change personal information */
 router.get('/account/details', (req, res) => {
     //if the session is logged in, render the account details page
     var session = req.session;
+    var user = session.username;
+    user.split(/-.*_/)
     if (session != null) {
-        db.collection('users').findOne({ username: session.username }, function (err, result) {
-            res.render('./pages/details.ejs', { user: result });
-        })
+        if (user.length == 2) {
+            db.collection('users').findOne({ fname: user[0], lname: user[1] }, function (err, result) {
+                res.render('./pages/details.ejs', { user: result });
+            })
+        }
+        else {
+            db.collection('users').find({ fname: user[0], lname: user[1] }, function (err, result) {
+                res.render('./pages/details.ejs', { user: result[user[2]] });
+            })
+        }
     }
 })
 /*Edit/change account information*/
 router.get('/acount/settings', (req, res) => {
     //if the session is logged in, render the account settings page
     var session = req.session;
+    var user = session.username;
+    user.split(/-.*_/)
     if (session != null) {
-        db.collection('users').findOne({ username: session.username }, function (err, result) {
-            res.render('./pages/settings.ejs', { user: result });
-        })
+        if (user.length == 2) {
+            db.collection('users').findOne({ fname: user[0], lname: user[1] }, function (err, result) {
+                res.render('./pages/settings.ejs', { user: result });
+            })
+        }
+        else {
+            db.collection('users').find({ fname: user[0], lname: user[1] }, function (err, result) {
+                res.render('./pages/settings.ejs', { user: result[user[2]] });
+            })
+        }
     }
 })
 /*Authored podcasts*/
-router.get('/users/:first-name-last-name/podcasts/authored', (req, res) => {
+router.get('/users/:first_name_last_name/podcasts/authored', (req, res) => {
     var name = req.params.first_name_last_name;
     name.split('-')
-    db.collection('users').find({ fname: name[0] }, function (err, result) {
+    db.collection('users').find({ fname: name[0], lname: name[1] }, function (err, result) {
         if (err) {
             console.log('Error finding users');
             throw err;
         }
-        forEach(result, function (user) {
-            if (user['lname'] == name[1])
-                db.collection('podcasts').find({ author: user['username'] }).toArray(function (err, result2) {
-                    if (err) {
-                        console.log('Error finding podcasts');
-                        throw err;
-                    }
-                    res.send(result2);
-                })
-        })
+        res.render('authored.ejs', { user: result });
 
     })
 })
@@ -173,21 +199,17 @@ router.get('/users/:first-name-last-name/podcasts/authored', (req, res) => {
 router.get('/users/:first_name_last_name/podcasts/saved', (req, res) => {
     var name = req.params.first_name_last_name;
     name.split('-')
-    db.collection('users').find({ fname: name[0] }, function (err, result) {
+    db.collection('users').find({ fname: name[0], lname: name[1] }, function (err, result) {
         if (err) {
             console.log('Error finding users');
             throw err;
         }
-        forEach(result, function (user) {
-            if (user['lname'] == name[1])
-                db.collection('podcasts').find({ saved: user['username'] }).toArray(function (err, result2) {
-                    if (err) {
-                        console.log('Error finding podcasts');
-                        throw err;
-                    }
-                    res.send(result2);
-                })
-        })
+        if (name.length == 2) {
+            res.render('saved.ejs', { user: result });
+        }
+        else {
+            res.render('saved.ejs', { user: result[name[2]] });
+        }
     })
 })
 /* Podcast upload page */
@@ -204,13 +226,28 @@ router.get('/podcasts/create', (req, res) => {
 router.get('/podcasts/:podcast_title/edit', (req, res) => {
     var title = req.params.podcast_title;
     var session = req.session;
+    var name = session.username;
+    name.split('-')
+    var podcast = [];
+    name.split('-')
     db.collection('podcasts').findOne({ title: title }, function (err, result) {
         if (err) {
             console.log('Error finding podcasts');
             throw err;
         }
-        if (session.role == 'admin' || (session.role == 'author' && result.author == session.username)) {
-            res.render('./pages/podcasts/edit.ejs', { podcast: result });
+        if (name.length == 2) {
+            if (session.role == 'admin' || (session.role == 'author' && result.fname == name[0] && result.lname == name[1])) {
+
+                res.render('./pages/podcasts/edit.ejs', { podcast: result });
+            }
+        }
+        else {
+            podcast = result;
+        }
+    })
+    db.collection('users').find({ fname: name[0], lname: name[1] }, function (err, result) {
+        if (podcast in result[name[2]].authored) {
+            res.render('./pages/podcasts/edit.ejs', { podcast: podcast });
         }
     })
 })
@@ -250,6 +287,11 @@ router.get('/api/search/date/:date', function (req, res) {
 /* Home page (logged in) */
 router.get('/api', function (req, res) {
     db.collection('podcasts').find().sort({ _id: -1 }).limit(10).toArray(function (err, result) {
+        if (err) {
+            console.log('Error finding podcasts');
+            throw err;
+        }
+        res.send(result);
     })
 })
 router.get('/api/search/keyword/:keyword', function (req, res) {
@@ -353,7 +395,14 @@ router.get('/api/keywords', function (req, res) {
         })
     })
 
-    forEach(keywords, function (keyword) { })
+    forEach(keywords, function (keyword) {
+        for (i = 0; i < sort.length; i++) {
+            if (sort[i][0] == keyword.name) {
+                send.push({ keyword: keyword.name, count: sort[i][1] });
+            }
+        }
+    })
+    res.send(send)
 })
 /* Tage page */
 router.get('/api/keywords/:keyword', function (req, res) {
@@ -411,16 +460,36 @@ router.patch('/api/podcasts/:podcast_title', function (req, res) {
 })
 router.delete('/api/podcasts/:podcast_title', function (req, res) {
     var podcast = req.params.podcast_title;
-    db.collection('podcasts').deleteOne({ title: podcast }, function (err, result) {
-        if (err) {
-            console.log('Error deleting podcast');
-            throw err;
+    var name = req.session.username;
+    var authored = [];
+    var podcaststuff = [];
+    name.split('-')
+    if (name != null) {
+        db.collection('users').find({ fname: name[0], lname: name[1] }).toArray(function (err, result) {
+            if (err) {
+                console.log('Error finding user');
+                throw err;
+            }
+            if (name > 2) {
+                authored = result[name[2]].authored;
+            }
+            authored = result[0].authored
+        })
+        db.collection('podcasts').findOne({ title: podcast }).toArray(function (err, result) {
+        if (podcast in authored) {
+            db.collection('podcasts').deleteOne({ title: podcast }, function (err, result) {
+                if (err) {
+                    console.log('Error deleting podcast');
+                    throw err;
+                }
+            })
         }
-        res.send(result);
-    })
+    }
 })
 router.patch('/api/podcasts/:podcast_title/actions/subscribe', function (req, res) {
-
+    var podcast = req.params.podcast_title;
+    db.collection('users').updateOne({ title: podcast }, { $set: { subscribed: true } }, function (err, result) {
+    })
 })
 router.patch('/api/podcasts/:podcast_title/actions/like', function (req, res) {
     var podcast = req.params.podcast_title;
@@ -495,13 +564,35 @@ router.get('/api/users/:first_name_last_name', function (req, res) {
 
 })
 router.post('/api/users/:first_name_last_name/actions/follow', function (req, res) {
+    var user = req.params.first_name_last_name;
 })
 router.get('/api/users/:first_name_last_name/podcasts/authored', function (req, res) {
+    var user = req.params.first_name_last_name;
+    db.collection('podcasts').find({ author: user }).toArray(function (err, result) { })
 })
 router.get('/api/users/:first_name_last_name/podcasts/saves', function (req, res) {
+    var user = req.params.first_name_last_name;
+    db.collection('podcasts').find({ saved: user }).toArray(function (err, result) { })
 })
 /* User/author account */
 router.get('/api/account', function (req, res) {
+    var user = req.body;
+    db.collection('users').findOne({ fname: user[0], lname: user[1] }, function (err, result) {
+        if (err) {
+            console.log('Error finding user');
+            throw err;
+        }
+        res.send(result);
+    })
 })
 router.patch('/api/account', function (req, res) {
+    var user = req.body;
+    db.collection('users').updateOne({ fname: user[0], lname: user[1] }, { $set: user }, function (err, result) {
+        if (err) {
+            console.log('Error updating user');
+            throw err;
+        }
+        res.send(result);
+    })
 })
+module.exports = { router };
